Fail if libpcsclite_nospy.so.1 already exists
git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@6050 0ce88b0d-b2fd-0310-8134-9614164e65ea

+++ src/spy/install_spy.sh
@@ -2,6 +2,12 @@
 
 cd /usr/lib
 
+if [ -f libpcsclite_nospy.so.1 ]
+then
+	echo "File libpcsclite_nospy.so.1 already exists"
+	exit
+fi
+
 # backup the real library
 cp libpcsclite.so.1 libpcsclite_nospy.so.1
